+++ test/connectionmonitor/SignalingAndMetricsConnectionMonitor.test.ts
@@ -15,8 +15,6 @@ import PingPongObserver from '../../src/pingpongobserver/PingPongObserver';
 import DefaultRealtimeController from '../../src/realtimecontroller/DefaultRealtimeController';
 import StatsCollector from '../../src/statscollector/StatsCollector';
 import { Maybe } from '../../src/utils/Types';
-import DefaultVideoTileController from '../../src/videotilecontroller/DefaultVideoTileController';
-import DefaultVideoTileFactory from '../../src/videotilefactory/DefaultVideoTileFactory';
 import DOMMockBuilder from '../dommock/DOMMockBuilder';
 
 // eslint-disable-next-line
@@ -26,7 +24,6 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
   let domMockBuilder: DOMMockBuilder;
   let audioVideoController: NoOpAudioVideoController;
   let realTimeController: DefaultRealtimeController;
-  let videoTileController: DefaultVideoTileController;
   let connectionHealthData: ConnectionHealthData;
   let pingPongStartCalled: boolean;
   let consecutiveMissedPongsCalled: boolean;
@@ -134,7 +131,6 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
   }
 
   const expect: Chai.ExpectStatic = chai.expect;
-  const assert: Chai.AssertStatic = chai.assert;
   let connectionMonitor: SignalingAndMetricsConnectionMonitor;
   let testClientMetricReport: TestClientMetricReport;
 
@@ -155,17 +151,11 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     setLastGoodSignalTimestampMsCalled = false;
     audioVideoController = new TestAudioVideoController();
     realTimeController = new TestRealtimeController(new NoOpMediaStreamBroker());
-    videoTileController = new DefaultVideoTileController(
-      new DefaultVideoTileFactory(),
-      audioVideoController,
-      new NoOpDebugLogger()
-    );
     connectionHealthData = new TestConnectionHealthData();
 
     connectionMonitor = new SignalingAndMetricsConnectionMonitor(
       audioVideoController,
       realTimeController,
-      videoTileController,
       connectionHealthData,
       new TestPingPong(),
       new StatsCollector(audioVideoController, new NoOpDebugLogger())
@@ -316,117 +306,5 @@ describe('SignalingAndMetricsConnectionMonitor', () => {
     connectionMonitor.stop();
     signalStrengthCallback(1);
     expect(setLastGoodSignalTimestampMsCalled).to.equal(false);
-  });
-
-  describe('video', () => {
-    it('can notify an observer when sending bandwidth changes', done => {
-      const bandwidth1 = 100;
-      const bandwidth2 = 1000;
-
-      class TestAudioVideoObserver implements AudioVideoObserver {
-        videoSendBandwidthDidChange(
-          currentSendBandwidthKbps: number,
-          previousSendBandwidthKbps: number
-        ): void {
-          expect(currentSendBandwidthKbps).to.equal(bandwidth2 / 1000);
-          expect(previousSendBandwidthKbps).to.equal(bandwidth1 / 1000);
-          done();
-        }
-      }
-
-      audioVideoController.addObserver(new TestAudioVideoObserver());
-      testClientMetricReport.availableOutgoingBitrate = bandwidth1;
-      sendClientMetricReport(testClientMetricReport);
-      testClientMetricReport.availableOutgoingBitrate = bandwidth2;
-      sendClientMetricReport(testClientMetricReport);
-    });
-
-    it('can notify an observer when receiving bandwidth changes', done => {
-      const bandwidth1 = 100;
-      const bandwidth2 = 1000;
-
-      class TestAudioVideoObserver implements AudioVideoObserver {
-        videoReceiveBandwidthDidChange(
-          currentRecvBandwidthKbps: number,
-          previousRecvBandwidthKbps: number
-        ): void {
-          expect(currentRecvBandwidthKbps).to.equal(bandwidth2 / 1000);
-          expect(previousRecvBandwidthKbps).to.equal(bandwidth1 / 1000);
-          done();
-        }
-      }
-
-      audioVideoController.addObserver(new TestAudioVideoObserver());
-      testClientMetricReport.availableIncomingBitrate = bandwidth1;
-      sendClientMetricReport(testClientMetricReport);
-      testClientMetricReport.availableIncomingBitrate = bandwidth2;
-      sendClientMetricReport(testClientMetricReport);
-    });
-
-    it('can notify an observer of the monitor video uplink state', done => {
-      const bitrate = 100;
-      const packets = 1000;
-
-      class TestAudioVideoObserver implements AudioVideoObserver {
-        videoSendHealthDidChange(
-          videoUpstreamBitrateKbps: number,
-          videoUpstreamPacketPerSecond: number
-        ): void {
-          expect(videoUpstreamBitrateKbps).to.equal(bitrate / 1000);
-          expect(videoUpstreamPacketPerSecond).to.equal(packets);
-          done();
-        }
-      }
-
-      videoTileController.startLocalVideoTile();
-      audioVideoController.addObserver(new TestAudioVideoObserver());
-      testClientMetricReport.videoUpstreamBitrate = bitrate;
-      testClientMetricReport.videoPacketSentPerSecond = packets;
-      sendClientMetricReport(testClientMetricReport);
-    });
-
-    it('does not notify an observer if the raw metrics value is not a number', () => {
-      class TestAudioVideoObserver implements AudioVideoObserver {
-        videoSendBandwidthDidChange(
-          _currentSendBandwidthKbps: number,
-          _previousSendBandwidthKbps: number
-        ): void {
-          assert.fail();
-        }
-
-        videoReceiveBandwidthDidChange(
-          _currentRecvBandwidthKbps: number,
-          _previousRecvBandwidthKbps: number
-        ): void {
-          assert.fail();
-        }
-      }
-
-      audioVideoController.addObserver(new TestAudioVideoObserver());
-      testClientMetricReport.availableOutgoingBitrate = 100;
-      testClientMetricReport.availableIncomingBitrate = 1000;
-      sendClientMetricReport(testClientMetricReport);
-      testClientMetricReport.availableOutgoingBitrate = '100';
-      testClientMetricReport.availableIncomingBitrate = '1000';
-      sendClientMetricReport(testClientMetricReport);
-    });
-
-    it('can notify an observer but always sends 0 if the raw video upstream bitrate is not a number', done => {
-      class TestAudioVideoObserver implements AudioVideoObserver {
-        videoSendHealthDidChange(
-          videoUpstreamBitrateKbps: number,
-          _videoUpstreamPacketPerSecond: number
-        ): void {
-          expect(videoUpstreamBitrateKbps).to.equal(0);
-          done();
-        }
-      }
-
-      videoTileController.startLocalVideoTile();
-      audioVideoController.addObserver(new TestAudioVideoObserver());
-      testClientMetricReport.videoUpstreamBitrate = '100';
-      testClientMetricReport.videoPacketSentPerSecond = 1000;
-      sendClientMetricReport(testClientMetricReport);
-    });
   });
 });